+++ EngineSIU/EngineSIU/Engine/Source/Contents/Actors/GoalPlatformActor.cpp
@@ -9,6 +9,6 @@ AGoalPlatformActor::AGoalPlatformActor()
     RootComponent = BoxComponent;
 
     MeshComponent = AddComponent<UStaticMeshComponent>(FName("MeshComponent_0"));
-    MeshComponent->SetStaticMesh(FObjManager::GetStaticMesh(L"Contents/FryBasket/FryBasket.obj"));
+    MeshComponent->SetStaticMesh(FResourceManager::GetStaticMesh(L"Contents/FryBasket/FryBasket.obj"));
     MeshComponent->SetupAttachment(BoxComponent);
 }
