+++ EngineSIU/EngineSIU/Engine/Source/Contents/Actors/TriggerBox.cpp
@@ -1,4 +1,4 @@
-﻿#include "TriggerBox.h"
+#include "TriggerBox.h"
 #include "Components/StaticMeshComponent.h"
 #include "Components/BoxComponent.h"
 #include "Engine/FObjLoader.h"
@@ -9,7 +9,7 @@ ATriggerBox::ATriggerBox()
     RootComponent = BoxComponent;
 
     MeshComponent = AddComponent<UStaticMeshComponent>(FName("MeshComponent_0"));
-    MeshComponent->SetStaticMesh(FObjManager::GetStaticMesh(L"Contents/FryBasket/FryBasket.obj"));
+    MeshComponent->SetStaticMesh(FResourceManager::GetStaticMesh(L"Contents/FryBasket/FryBasket.obj"));
     MeshComponent->SetupAttachment(BoxComponent);
 }
