+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Actors/FireballActor.cpp
@@ -9,12 +9,12 @@
 
 AFireballActor::AFireballActor()
 {
-    FObjManager::CreateStaticMesh("Contents/Sphere.obj");
+    FResourceManager::CreateStaticMesh("Contents/Sphere.obj");
 
 
     SphereComp = AddComponent<USphereComp>("USphereComp_0");
     
-    SphereComp->SetStaticMesh(FObjManager::GetStaticMesh(L"Contents/Sphere.obj"));
+    SphereComp->SetStaticMesh(FResourceManager::GetStaticMesh(L"Contents/Sphere.obj"));
   
     PointLightComponent = AddComponent<UPointLightComponent>("UPointLightComponent_0");
