+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Components/CubeComp.cpp
@@ -19,8 +19,8 @@ void UCubeComp::InitializeComponent()
     //SetStaticMesh(FManagerGetStaticMesh(L"helloBlender.obj"));
     // 
     // Begin Test
-    FObjManager::CreateStaticMesh("Contents/Reference/Reference.obj");
-    SetStaticMesh(FObjManager::GetStaticMesh(L"Reference.obj"));
+    FResourceManager::CreateStaticMesh("Contents/Reference/Reference.obj");
+    SetStaticMesh(FResourceManager::GetStaticMesh(L"Reference.obj"));
     // End Test
 }
