+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Engine/Asset/SkeletalMeshAsset.h
@@ -32,4 +32,6 @@ struct FSkeletalMeshRenderData
 
     FVector BoundingBoxMin;
     FVector BoundingBoxMax;
+
+    TArray<FVertexSkeletal> SourceVertices;
 };
