+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Engine/AssetManager.cpp
@@ -59,9 +59,9 @@ void UAssetManager::LoadObjFiles()
             AssetRegistry->PathNameToAssetInfo.Add(NewAssetInfo.AssetName, NewAssetInfo);
             
             FString MeshName = NewAssetInfo.PackagePath.ToString() + "/" + NewAssetInfo.AssetName.ToString();
-            FObjManager::CreateStaticMesh(MeshName);
+            FResourceManager::CreateStaticMesh(MeshName);
             // ObjFileNames.push_back(UGTLStringLibrary::StringToWString(Entry.path().string()));
-            // FObjManager::LoadObjStaticMeshAsset(UGTLStringLibrary::StringToWString(Entry.path().string()));
+            // FResourceManager::LoadObjStaticMeshAsset(UGTLStringLibrary::StringToWString(Entry.path().string()));
         }
     }
 }
