+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Engine/FObjLoader.cpp
@@ -634,295 +634,3 @@ void FObjLoader::CalculateTangent(FStaticMeshVertex& PivotVertex, const FStaticM
     PivotVertex.TangentW = Sign;
 }
 
-FStaticMeshRenderData* FObjManager::LoadObjStaticMeshAsset(const FString& PathFileName)
-{
-    FStaticMeshRenderData* NewStaticMesh = new FStaticMeshRenderData();
-
-    if ( const auto It = ObjStaticMeshMap.Find(PathFileName))
-    {
-        return *It;
-    }
-
-    FWString BinaryPath = (PathFileName + ".bin").ToWideString();
-    if (std::ifstream(BinaryPath).good())
-    {
-        if (LoadStaticMeshFromBinary(BinaryPath, *NewStaticMesh))
-        {
-            ObjStaticMeshMap.Add(PathFileName, NewStaticMesh);
-            return NewStaticMesh;
-        }
-    }
-
-    // Parse OBJ
-    FObjInfo NewObjInfo;
-    bool Result = FObjLoader::ParseOBJ(PathFileName, NewObjInfo);
-
-    if (!Result)
-    {
-        delete NewStaticMesh;
-        return nullptr;
-    }
-
-    // Material
-    if (NewObjInfo.MaterialSubsets.Num() > 0)
-    {
-        Result = FObjLoader::ParseMaterial(NewObjInfo, *NewStaticMesh);
-
-        if (!Result)
-        {
-            delete NewStaticMesh;
-            return nullptr;
-        }
-
-        CombineMaterialIndex(*NewStaticMesh);
-
-        for (int materialIndex = 0; materialIndex < NewStaticMesh->Materials.Num(); materialIndex++) {
-            CreateMaterial(NewStaticMesh->Materials[materialIndex]);
-        }
-    }
-
-    // Convert FStaticMeshRenderData
-    Result = FObjLoader::ConvertToStaticMesh(NewObjInfo, *NewStaticMesh);
-    if (!Result)
-    {
-        delete NewStaticMesh;
-        return nullptr;
-    }
-
-    SaveStaticMeshToBinary(BinaryPath, *NewStaticMesh); 
-    ObjStaticMeshMap.Add(PathFileName, NewStaticMesh);
-    return NewStaticMesh;
-}
-
-void FObjManager::CombineMaterialIndex(FStaticMeshRenderData& OutFStaticMesh)
-{
-    for (int32 i = 0; i < OutFStaticMesh.MaterialSubsets.Num(); i++)
-    {
-        FString MatName = OutFStaticMesh.MaterialSubsets[i].MaterialName;
-        for (int32 j = 0; j < OutFStaticMesh.Materials.Num(); j++)
-        {
-            if (OutFStaticMesh.Materials[j].MaterialName == MatName)
-            {
-                OutFStaticMesh.MaterialSubsets[i].MaterialIndex = j;
-                break;
-            }
-        }
-    }
-}
-
-bool FObjManager::SaveStaticMeshToBinary(const FWString& FilePath, const FStaticMeshRenderData& StaticMesh)
-{
-    std::ofstream File(FilePath, std::ios::binary);
-    if (!File.is_open())
-    {
-        assert("CAN'T SAVE STATIC MESH BINARY FILE");
-        return false;
-    }
-
-    // Object Name
-    Serializer::WriteFWString(File, StaticMesh.ObjectName);
-
-    // Display Name
-    Serializer::WriteFString(File, StaticMesh.DisplayName);
-
-    // Vertices
-    uint32 VertexCount = StaticMesh.Vertices.Num();
-    File.write(reinterpret_cast<const char*>(&VertexCount), sizeof(VertexCount));
-    File.write(reinterpret_cast<const char*>(StaticMesh.Vertices.GetData()), VertexCount * sizeof(FStaticMeshVertex));
-
-    // Indices
-    uint32 IndexCount = StaticMesh.Indices.Num();
-    File.write(reinterpret_cast<const char*>(&IndexCount), sizeof(IndexCount));
-    File.write(reinterpret_cast<const char*>(StaticMesh.Indices.GetData()), IndexCount * sizeof(UINT));
-
-    // Materials
-    uint32 MaterialCount = StaticMesh.Materials.Num();
-    File.write(reinterpret_cast<const char*>(&MaterialCount), sizeof(MaterialCount));
-    for (const FObjMaterialInfo& Material : StaticMesh.Materials)
-    {
-        Serializer::WriteFString(File, Material.MaterialName);
-        
-        File.write(reinterpret_cast<const char*>(&Material.TextureFlag), sizeof(Material.TextureFlag));
-        
-        File.write(reinterpret_cast<const char*>(&Material.bTransparent), sizeof(Material.bTransparent));
-        File.write(reinterpret_cast<const char*>(&Material.DiffuseColor), sizeof(Material.DiffuseColor));
-        File.write(reinterpret_cast<const char*>(&Material.SpecularColor), sizeof(Material.SpecularColor));
-        File.write(reinterpret_cast<const char*>(&Material.AmbientColor), sizeof(Material.AmbientColor));
-        File.write(reinterpret_cast<const char*>(&Material.EmissiveColor), sizeof(Material.EmissiveColor));
-        
-        File.write(reinterpret_cast<const char*>(&Material.Shininess), sizeof(Material.Shininess));
-        File.write(reinterpret_cast<const char*>(&Material.IOR), sizeof(Material.IOR));
-        File.write(reinterpret_cast<const char*>(&Material.Transparency), sizeof(Material.Transparency));
-        File.write(reinterpret_cast<const char*>(&Material.BumpMultiplier), sizeof(Material.BumpMultiplier));
-        File.write(reinterpret_cast<const char*>(&Material.IlluminanceModel), sizeof(Material.IlluminanceModel));
-
-        File.write(reinterpret_cast<const char*>(&Material.Metallic), sizeof(Material.Metallic));
-        File.write(reinterpret_cast<const char*>(&Material.Roughness), sizeof(Material.Roughness));
-
-        for (uint8 i = 0; i < static_cast<uint8>(EMaterialTextureSlots::MTS_MAX); ++i)
-        {
-            Serializer::WriteFString(File, Material.TextureInfos[i].TextureName);
-            Serializer::WriteFWString(File, Material.TextureInfos[i].TexturePath);
-            File.write(reinterpret_cast<const char*>(&Material.TextureInfos[i].bIsSRGB), sizeof(Material.TextureInfos[i].bIsSRGB));
-        }
-    }
-
-    // Material Subsets
-    uint32 SubsetCount = StaticMesh.MaterialSubsets.Num();
-    File.write(reinterpret_cast<const char*>(&SubsetCount), sizeof(SubsetCount));
-    for (const FMaterialSubset& Subset : StaticMesh.MaterialSubsets)
-    {
-        Serializer::WriteFString(File, Subset.MaterialName);
-        File.write(reinterpret_cast<const char*>(&Subset.IndexStart), sizeof(Subset.IndexStart));
-        File.write(reinterpret_cast<const char*>(&Subset.IndexCount), sizeof(Subset.IndexCount));
-        File.write(reinterpret_cast<const char*>(&Subset.MaterialIndex), sizeof(Subset.MaterialIndex));
-    }
-
-    // Bounding Box
-    File.write(reinterpret_cast<const char*>(&StaticMesh.BoundingBoxMin), sizeof(FVector));
-    File.write(reinterpret_cast<const char*>(&StaticMesh.BoundingBoxMax), sizeof(FVector));
-
-    File.close();
-    return true;
-}
-
-bool FObjManager::LoadStaticMeshFromBinary(const FWString& FilePath, FStaticMeshRenderData& OutStaticMesh)
-{
-    std::ifstream File(FilePath, std::ios::binary);
-    if (!File.is_open())
-    {
-        assert("CAN'T OPEN STATIC MESH BINARY FILE");
-        return false;
-    }
-
-    TArray<TPair<FWString, bool>> Textures;
-
-    // Object Name
-    Serializer::ReadFWString(File, OutStaticMesh.ObjectName);
-
-    //// Path Name
-    //Serializer::ReadFWString(File, OutStaticMesh.PathName);
-
-    // Display Name
-    Serializer::ReadFString(File, OutStaticMesh.DisplayName);
-
-    // Vertices
-    uint32 VertexCount = 0;
-    File.read(reinterpret_cast<char*>(&VertexCount), sizeof(VertexCount));
-    OutStaticMesh.Vertices.SetNum(VertexCount);
-    File.read(reinterpret_cast<char*>(OutStaticMesh.Vertices.GetData()), VertexCount * sizeof(FStaticMeshVertex));
-
-    // Indices
-    uint32 IndexCount = 0;
-    File.read(reinterpret_cast<char*>(&IndexCount), sizeof(IndexCount));
-    OutStaticMesh.Indices.SetNum(IndexCount);
-    File.read(reinterpret_cast<char*>(OutStaticMesh.Indices.GetData()), IndexCount * sizeof(UINT));
-
-    // Material
-    uint32 MaterialCount = 0;
-    File.read(reinterpret_cast<char*>(&MaterialCount), sizeof(MaterialCount));
-    OutStaticMesh.Materials.SetNum(MaterialCount);
-    for (FObjMaterialInfo& Material : OutStaticMesh.Materials)
-    {
-        Serializer::ReadFString(File, Material.MaterialName);
-        File.read(reinterpret_cast<char*>(&Material.TextureFlag), sizeof(Material.TextureFlag));
-        
-        File.read(reinterpret_cast<char*>(&Material.bTransparent), sizeof(Material.bTransparent));
-        File.read(reinterpret_cast<char*>(&Material.DiffuseColor), sizeof(Material.DiffuseColor));
-        File.read(reinterpret_cast<char*>(&Material.SpecularColor), sizeof(Material.SpecularColor));
-        File.read(reinterpret_cast<char*>(&Material.AmbientColor), sizeof(Material.AmbientColor));
-        File.read(reinterpret_cast<char*>(&Material.EmissiveColor), sizeof(Material.EmissiveColor));
-        
-        File.read(reinterpret_cast<char*>(&Material.Shininess), sizeof(Material.Shininess));
-        File.read(reinterpret_cast<char*>(&Material.IOR), sizeof(Material.IOR));
-        File.read(reinterpret_cast<char*>(&Material.Transparency), sizeof(Material.Transparency));
-        File.read(reinterpret_cast<char*>(&Material.BumpMultiplier), sizeof(Material.BumpMultiplier));
-        File.read(reinterpret_cast<char*>(&Material.IlluminanceModel), sizeof(Material.IlluminanceModel));
-
-        File.read(reinterpret_cast<char*>(&Material.Metallic), sizeof(Material.Metallic));
-        File.read(reinterpret_cast<char*>(&Material.Roughness), sizeof(Material.Roughness));
-
-        uint8 TextureNum = static_cast<uint8>(EMaterialTextureSlots::MTS_MAX);
-        Material.TextureInfos.SetNum(TextureNum);
-        for (uint8 i = 0; i < TextureNum; ++i)
-        {
-            Serializer::ReadFString(File, Material.TextureInfos[i].TextureName);
-            Serializer::ReadFWString(File, Material.TextureInfos[i].TexturePath);
-            File.read(reinterpret_cast<char*>(&Material.TextureInfos[i].bIsSRGB), sizeof(Material.TextureInfos[i].bIsSRGB));
-
-            Textures.AddUnique({Material.TextureInfos[i].TexturePath, Material.TextureInfos[i].bIsSRGB});
-        }
-    }
-
-    // Material Subset
-    uint32 SubsetCount = 0;
-    File.read(reinterpret_cast<char*>(&SubsetCount), sizeof(SubsetCount));
-    OutStaticMesh.MaterialSubsets.SetNum(SubsetCount);
-    for (FMaterialSubset& Subset : OutStaticMesh.MaterialSubsets)
-    {
-        Serializer::ReadFString(File, Subset.MaterialName);
-        File.read(reinterpret_cast<char*>(&Subset.IndexStart), sizeof(Subset.IndexStart));
-        File.read(reinterpret_cast<char*>(&Subset.IndexCount), sizeof(Subset.IndexCount));
-        File.read(reinterpret_cast<char*>(&Subset.MaterialIndex), sizeof(Subset.MaterialIndex));
-    }
-
-    // Bounding Box
-    File.read(reinterpret_cast<char*>(&OutStaticMesh.BoundingBoxMin), sizeof(FVector));
-    File.read(reinterpret_cast<char*>(&OutStaticMesh.BoundingBoxMax), sizeof(FVector));
-
-    File.close();
-
-    // Texture Load
-    if (Textures.Num() > 0)
-    {
-        for (const TPair<FWString, bool>& Texture : Textures)
-        {
-            if (FEngineLoop::ResourceManager.GetTexture(Texture.Key) == nullptr)
-            {
-                FEngineLoop::ResourceManager.LoadTextureFromFile(FEngineLoop::GraphicDevice.Device, Texture.Key.c_str(), Texture.Value);
-            }
-        }
-    }
-
-    return true;
-}
-
-UMaterial* FObjManager::CreateMaterial(FObjMaterialInfo materialInfo)
-{
-    if (MaterialMap[materialInfo.MaterialName] != nullptr)
-        return MaterialMap[materialInfo.MaterialName];
-
-    UMaterial* newMaterial = FObjectFactory::ConstructObject<UMaterial>(nullptr); // Material은 Outer가 없이 따로 관리되는 객체이므로 Outer가 없음으로 설정. 추후 Garbage Collection이 추가되면 AssetManager를 생성해서 관리.
-    newMaterial->SetMaterialInfo(materialInfo);
-    MaterialMap.Add(materialInfo.MaterialName, newMaterial);
-    return newMaterial;
-}
-
-UMaterial* FObjManager::GetMaterial(FString name)
-{
-    return MaterialMap[name];
-}
-
-UStaticMesh* FObjManager::CreateStaticMesh(const FString& filePath)
-{
-    FStaticMeshRenderData* StaticMeshRenderData = FObjManager::LoadObjStaticMeshAsset(filePath);
-
-    if (StaticMeshRenderData == nullptr) return nullptr;
-
-    UStaticMesh* StaticMesh = GetStaticMesh(StaticMeshRenderData->ObjectName);
-    if (StaticMesh != nullptr)
-    {
-        return StaticMesh;
-    }
-
-    StaticMesh = FObjectFactory::ConstructObject<UStaticMesh>(nullptr); // TODO: 추후 AssetManager를 생성해서 관리.
-    StaticMesh->SetData(StaticMeshRenderData);
-
-    StaticMeshMap.Add(StaticMeshRenderData->ObjectName, StaticMesh); // TODO: 장기적으로 보면 파일 이름 대신 경로를 Key로 사용하는게 좋음.
-    return StaticMesh;
-}
-
-UStaticMesh* FObjManager::GetStaticMesh(FWString name)
-{
-    return StaticMeshMap[name];
-}
