--- EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Engine/ResourceMgr.h
@@ -1,20 +1,0 @@
-#pragma once
-#include <memory>
-#include "Texture.h"
-#include "Container/Map.h"
-
-class FRenderer;
-class FGraphicsDevice;
-class FResourceMgr
-{
-
-public:
-    void Initialize(FRenderer* renderer, FGraphicsDevice* device);
-    void Release(FRenderer* renderer);
-    HRESULT LoadTextureFromFile(ID3D11Device* device, const wchar_t* filename, bool bIsSRGB = true);
-    HRESULT LoadTextureFromDDS(ID3D11Device* device, ID3D11DeviceContext* context, const wchar_t* filename);
-
-    std::shared_ptr<FTexture> GetTexture(const FWString& name) const;
-private:
-    TMap<FWString, std::shared_ptr<FTexture>> textureMap;
-};
